+++ MobileClients/Android/nervousnetLIB/src/ch/ethz/coss/nervousnet/lib/GyroReading.java
@@ -1,10 +1,8 @@
 
 package ch.ethz.coss.nervousnet.lib;
 
-import android.os.BatteryManager;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.util.Log;
 
 /**
  * @author prasad
@@ -13,7 +11,7 @@ public class GyroReading extends SensorReading {
 
 	private float[] values = new float[3];
 
-	public GyroReading(int timestamp, float[] values) {
+	public GyroReading(long timestamp, float[] values) {
 		this.timestamp = timestamp;
 		this.values = values;
 	}
@@ -27,7 +25,7 @@ public class GyroReading extends SensorReading {
 
 	public void readFromParcel(Parcel in) {
 
-		timestamp = in.readInt();
+		timestamp = in.readLong();
 		in.readFloatArray(values);
 	}
 
@@ -61,7 +59,7 @@ public class GyroReading extends SensorReading {
 	 */
 	@Override
 	public void writeToParcel(Parcel out, int flags) {
-		out.writeInt(timestamp);
+		out.writeLong(timestamp);
 		out.writeFloatArray(values);
 
 	}
